[deckhouse] Refactor package repository operation service
