+++ deckhouse-controller/pkg/apis/deckhouse.io/v1alpha1/package_repository_operation.go
@@ -45,6 +45,11 @@ const (
 	PackagesRepositoryOperationLabelOperationType                                = "packages.deckhouse.io/operation-type"
 	PackageRepositoryOperationTypeUpdate          PackageRepositoryOperationType = "Update"
 
+	// PackageTypeLabel is a label on Docker images that indicates the package type
+	PackagesRepositoryOperationLabelPackageType = "io.deckhouse.package.type"
+	PackageTypeApplication                      = "Application"
+	PackageTypeModule                           = "Module"
+
 	PackagesRepositoryOperationLabelOperationTrigger = "packages.deckhouse.io/operation-trigger"
 	PackagesRepositoryTriggerManual                  = "manual"
 	PackagesRepositoryTriggerAuto                    = "auto"
